requirements.txt
python-telegram-bot==21.6
anthropic==0.49.0
fpdf2>=2.7.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
